requirements.txt
sentry>=4.3,<4.4
mysql-python>=1.2,<1.3
# For gevent support with sentry:
gevent>=0.13,<0.14
# For udp support in sentry:
eventlet>=0.9,<0.10
# For queue/buffers support in sentry:
redis>=2.4,<2.5
nydus>=0.8,<0.9
# For nice gunicorn/celery process names:
setproctitle>=1.1,<1.2
# A middleware to force SSL on sentry
django-sslify==0.2
